feat(follow/unfollow): Implement follow/unfollow in user model
- Add a following(manytomanyfield) field in the User model
- Move follow/unfollow serializers and views to the auth app
- Refactor follow/unfollow code to utilize user model

[Finishes #163388142
]
