+++ authors/apps/authentication/models.py
@@ -88,6 +88,9 @@ class User(AbstractBaseUser, PermissionsMixin):
     # A timestamp reprensenting when this object was last updated.
     updated_at = models.DateTimeField(auto_now=True)
 
+    # Id for users a user is following
+    following = models.ManyToManyField('User', related_name='followers')
+
     # More fields required by Django when specifying a custom user model.
 
     # The `USERNAME_FIELD` property tells us which field we will use to log in.
