--- authors/apps/friends/admin.py
@@ -1,1 +1,0 @@
-from django.contrib import admin
