--- authors/apps/friends/apps.py
@@ -1,5 +1,0 @@
-from django.apps import AppConfig
-
-
-class FriendsConfig(AppConfig):
-    name = 'friends'
