+++ authors/settings.py
@@ -46,8 +46,7 @@ INSTALLED_APPS = [
     'authors.apps.authentication',
     'authors.apps.core',
     'authors.apps.profiles',
-    'authors.apps.base',
-    'authors.apps.friends'
+    'authors.apps.base'
 ]
 
 MIDDLEWARE = [
