+++ authors/urls.py
@@ -22,7 +22,5 @@ urlpatterns = [
     path('', include('authors.apps.base.urls')),
     path('api/v1/docs/', schema_view),
     path('api/v1/', include(('authors.apps.authentication.urls',
-                             'authentication'), namespace='authentication')),
-    path('api/v1/', include(('authors.apps.friends.urls',
-                             'following'), namespace='following'))
+                             'authentication'), namespace='authentication'))
 ]
